Fix tests inheritance
